Added arch support
