+++ apps/mbus/Makefile.mbus
@@ -1,1 +1,1 @@
-mbus_src = mbus.c ringbuf-mbus.c usart1.c mbus-supported-devices.c
+mbus_src = mbus.c mbus-supported-devices.c
