+++ examples/avr-rss2/mbus/Makefile
@@ -10,7 +10,9 @@ PRINTF_LIB_MIN  = -Wl,-u,vfprintf -lprintf_min
 PRINTF_LIB      = $(PRINTF_LIB_FLT)
 CLIBS           = $(PRINTF_LIB)
 
-#MODULES += dev/bme280
+MODULES += dev/mbus
+
+#PROJECT_OBJECTFILES += dev/mbus-arch.o
 
 CUSTOM_RULE_LINK = 1
